cortex-medical-hf/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
transformers==4.36.2
torch==2.1.2
sentence-transformers==2.2.2
scikit-learn==1.3.2
faiss-cpu==1.7.4
pandas==2.1.4
numpy==1.25.2
pyarrow==14.0.2
pydantic==2.5.2
python-multipart==0.0.6